Refactored locale file to load based on ENV vars
