+++ src/pages/over.tsx
@@ -3,12 +3,23 @@ import Head from 'next/head';
 import Layout, { FunctionComponentWithLayout } from 'components/layout';
 import MaxWidth from 'components/maxWidth';
 
-import text from 'locale/nl.json';
+import text from 'locale';
 import styles from './over.module.scss';
 import ReplaceLinks from 'components/replaceLinks';
 
 import openGraphImage from 'assets/sharing/og-over.png?url';
 import twitterImage from 'assets/sharing/twitter-over.png?url';
+
+interface IVraagEnAntwoord {
+  vraag: {
+    translation: string;
+    notes: string;
+  };
+  antwoord: {
+    translation: string;
+    notes: string;
+  };
+}
 
 const Over: FunctionComponentWithLayout = () => {
   return (
@@ -36,14 +47,16 @@ const Over: FunctionComponentWithLayout = () => {
             <p>{text.over_disclaimer.text.translation}</p>
             <h2>{text.over_veelgestelde_vragen.text.translation}</h2>
             <dl className={styles.faqList}>
-              {text.over_veelgestelde_vragen.vragen.map((item) => (
-                <>
-                  <dt>{item.vraag.translation}</dt>
-                  <dd>
-                    <ReplaceLinks>{item.antwoord.translation}</ReplaceLinks>
-                  </dd>
-                </>
-              ))}
+              {text.over_veelgestelde_vragen.vragen.map(
+                (item: IVraagEnAntwoord) => (
+                  <>
+                    <dt>{item.vraag.translation}</dt>
+                    <dd>
+                      <ReplaceLinks>{item.antwoord.translation}</ReplaceLinks>
+                    </dd>
+                  </>
+                )
+              )}
             </dl>
           </div>
         </MaxWidth>
