+++ src/pages/verantwoording.tsx
@@ -3,12 +3,23 @@ import Head from 'next/head';
 import Layout, { FunctionComponentWithLayout } from 'components/layout';
 import MaxWidth from 'components/maxWidth';
 
-import text from 'locale/nl.json';
+import text from 'locale';
 import styles from './over.module.scss';
 import ReplaceLinks from 'components/replaceLinks';
 
 import openGraphImage from 'assets/sharing/og-cijferverantwoording.png?url';
 import twitterImage from 'assets/sharing/twitter-cijferverantwoording.png?url';
+
+interface ICijfer {
+  cijfer: {
+    translation: string;
+    notes: string;
+  };
+  verantwoording: {
+    translation: string;
+    notes: string;
+  };
+}
 
 const Verantwoording: FunctionComponentWithLayout = () => {
   return (
@@ -32,7 +43,7 @@ const Verantwoording: FunctionComponentWithLayout = () => {
           <div className={styles.maxwidth}>
             <h2>{text.verantwoording.title.translation}</h2>
             <dl className={styles.faqList}>
-              {text.verantwoording.cijfers.map((item) => (
+              {text.verantwoording.cijfers.map((item: ICijfer) => (
                 <>
                   <dt>{item.cijfer.translation}</dt>
                   <dd>
